+++ app/src/main/java/cl/blackmind/app/NotesApp.kt
@@ -1,16 +1,15 @@
 package cl.blackmind.app
 
 import android.app.Application
-import com.jakewharton.threetenabp.AndroidThreeTen
 import cl.blackmind.app.di.appModule
 import cl.blackmind.app_api.di.appRemoteModule
 import cl.blackmind.app_domain.di.appDomainModule
 import cl.blackmind.app_local.di.appLocalModule
+import com.jakewharton.threetenabp.AndroidThreeTen
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
 import org.koin.core.logger.Level
-
 
 const val URL_BASE_EXPERIENCE_API = "BuildConfig.URL_BASE_EXPERIENCE_API"
 
@@ -32,7 +31,8 @@ open class NotesApp : Application() {
                     appDomainModule,
                     appLocalModule,
                     appRemoteModule
-                ))
+                )
+            )
         }
     }
 }
