+++ core-kotlin/build.gradle
@@ -5,11 +5,10 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.2'
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:${versions.coroutines_core}"
 
-    implementation "androidx.lifecycle:lifecycle-common:2.2.0"
+    implementation "androidx.lifecycle:lifecycle-common:${versions.androidx_lifecycle}"
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:${versions.androidx_livedata}"
-//    implementation "androidx.lifecycle:lifecycle-livedata:${versions.androidx_livedata}"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:${versions.androidx_lifecycle}"
     implementation "androidx.lifecycle:lifecycle-extensions:2.2.0"
 }
