+++ core/src/main/kotlin/cl/blackmind/core/base/responseWithMessage/ResponseWithMessage.kt
@@ -1,4 +1,4 @@
-package com.up.core.base.responseWithMessage
+package cl.blackmind.core.base.responseWithMessage
 
 data class ResponseWithMessage<T>(
     val data: T,
