+++ core/src/main/kotlin/cl/blackmind/core/coroutines/BaseUnitUseCase.kt
@@ -1,17 +1,17 @@
-package com.up.core.coroutines
+package cl.blackmind.core.coroutines
 
 import androidx.lifecycle.MutableLiveData
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.cancelChildren
 import kotlin.coroutines.CoroutineContext
 
-abstract class BaseUseCase<Q, W : MutableLiveData<*>>(
+abstract class BaseUnitUseCase<W : MutableLiveData<*>>(
     protected open val backgroundContext: CoroutineContext,
     protected open val foregroundContext: CoroutineContext
 ) {
     private var parentJob = Job()
 
-    abstract fun execute(liveData: W, params: Q)
+    abstract fun execute(liveData: W)
 
     protected fun newJob(): Job {
         parentJob = parentJob.run {
