+++ core/src/main/kotlin/cl/blackmind/core/coroutines/Result.kt
@@ -1,4 +1,4 @@
-package com.up.core.coroutines
+package cl.blackmind.core.coroutines
 
 sealed class Result<T> {
     data class OnSuccess<T>(val value: T) : Result<T>()
