+++ core/src/main/kotlin/cl/blackmind/core/extension/ConnectivityManager.kt
@@ -1,0 +1,2 @@
+package cl.blackmind.core.extension
+
