--- core/src/main/kotlin/com/up/core/extension/Anko.kt
@@ -1,2 +1,0 @@
-package com.up.core.extension
-
