+++ gradle-scripts/versions.gradle
@@ -25,6 +25,7 @@ ext {
             'circle_imageview'            : '3.0.0',
             'constraintlayout'            : '1.1.3',
             'coroutines'                  : '1.2.1',
+            'coroutines_core'             : '1.3.2',
             'core_ktx'                    : '1.2.0',
             'debug_db'                    : '1.0.6',
             'firebase_auth'               : '19.3.0',
